TimeInStateMerge
