+++ .idea/.idea.Game-Off-2022/.idea/indexLayout.xml
@@ -66,7 +66,6 @@
       <Path>.vscode</Path>
       <Path>Library</Path>
       <Path>Logs</Path>
-      <Path>Temp</Path>
       <Path>UserSettings</Path>
       <Path>obj</Path>
     </explicitExcludes>
