--- Assets/Game/Scripts/AIDecisionTimeInState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9c784a73dda7d649b0a18abdd9679e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
